+++ lib/src/internal/cit.dart
@@ -47,4 +47,5 @@ Future<MLoginCitResult> runCit(
     MLoginLog.error('CIT failed with unexpected Exception: $e\n$stacktrace');
     return MLoginCitResult.error(MLoginCitError.unknown);
   }
+  
 }