+++ lib/src/m_login.dart
@@ -277,8 +277,8 @@ class MLogin {
 
   ///
   /// Opens a page in the browser where the user can grant a SEPA mandate for
-  /// payments to the [payeeId] of the service, using the already entered bank
-  /// account information in the M-Login.
+  /// payments, using the already entered bank account information in the
+  /// M-Login.
   ///
   /// Calling this is usually the result of trying to trigger a payment for the
   /// given service ("checkout") and receiving a `mandate_missing` error message
@@ -292,8 +292,6 @@ class MLogin {
   /// Parameters, and where to get them:
   /// * The [methodId] identifies the SEPA payment method that needs a mandate.
   ///   If you tried to do a simple checkout, this is the default payment method
-  /// * The [payeeId] identifies the entity that shall receive the money. Talk
-  ///   to the M-Login team if you don't know what to put there
   ///
   /// Returns [true] in case the user finishes the page using the `done` button,
   /// [false] in any other case (e.g., the user pressed the "cancel" button in
@@ -304,15 +302,14 @@ class MLogin {
   /// was changed.
   ///
   Future<bool> openGrantSepaMandatePage(
-    String methodId,
-    String payeeId, {
+    String methodId, {
     bool ephemeral = false,
     String? overrideRedirectUri,
   }) {
     return openDataPage(
       this,
       portalUriSuffix: 'grantmandate',
-      extraParams: {'method_id': methodId, 'payee_id': payeeId},
+      extraParams: {'method_id': methodId},
       ephemeral: ephemeral,
       username: prefilledUsername,
       overrideRedirectUri: overrideRedirectUri,
@@ -325,12 +322,6 @@ class MLogin {
   /// Does not require previous Login. In case there is no valid login present
   /// in the browser, the user will be prompted to log in again.
   ///
-  /// The parameter [payeeId] is the id of the entity that will receive money,
-  /// when the user pays in this service using SEPA. Must be added here to
-  /// ensure that, when the user adds a new SEPA payment method, the correct
-  /// mandate is already requested as well (see documentation of
-  /// [openGrantSepaMandatePage]).
-  ///
   /// Returns [true] in case the user finishes the page using the `done` button,
   /// [false] in any other case (e.g., the user pressed the "cancel" button in
   /// the iOS browser, or the back button on Android)
@@ -339,15 +330,13 @@ class MLogin {
   /// or validated. It's safe to ignore the returned value and just assume data
   /// was changed.
   ///
-  Future<bool> openPaymentMethodsOverviewPage(
-    String payeeId, {
+  Future<bool> openPaymentMethodsOverviewPage({
     bool ephemeral = false,
     String? overrideRedirectUri,
   }) {
     return openDataPage(
       this,
       portalUriSuffix: 'paymentmethods',
-      extraParams: {'payee_id': payeeId},
       ephemeral: ephemeral,
       username: prefilledUsername,
       overrideRedirectUri: overrideRedirectUri,